Add ctors to the Runtime's builtin_ctx via InductiveTypeDef::ctor_context().
Previously, each of the constructors was being treated similarly to
primitive functions. They were hard-coded in a list constructed in the
builtins::builtins_ctx() function.

This commit generates the constructor values and types using the
InductiveType struct and its methods. They are still added to the
builtin_ctx of the Runtime, as they were before, but now they are
packaged properly with their types.