Typecheck match statements.
First, I change Runtime's inductive_typedefs from a Vec to a HashMap.
This allows me to quickly look up typedefs by their name.

Second, I pass the inductive_typedefs to the check_type function. This
is notable because typechecking requires more than the TypeContext now!
Any time a match statement is encountered, it needs to know the full
definition for the inductive type it's matching against.

When checking the type of a match statement, we dispatch to the new
check_type_match() helper method. This looks at the tag of the first
match arm and uses whatever constructor it finds to look up the
corresponding typedef:

If there is no corresponding typedef. (This means we have no notion of
default patterns).

Otherwise, we check the discriminee against against that type. Then, we
check against each match arm by calling the new check_type_match_arm()
function. This function takes the pattern and the argument types to the
constructor, asserts they are equal in number, zips them together into
an extended context, and then typechecks the body with respect to this
context.

Some notable issues:

    - the constructors are not yet checked for full coverage,
    - the constructors are not yet checked for redundancy, and
    - it is difficult to do a reverse-lookup from constructor to typedef