Change apply() and eval() to methods on Runtime.
Note that while doing this, I encountered a corner case of the
borrow-checker:

    https://github.com/rust-lang/rust/issues/59159

The following attribute suppresses the warning (which would otherwise
cause continuous integration to fail):

    #[allow(mutable_borrow_reservation_conflict)]

Taking a second look at this, it seems a bit dubious that the eval()
method should be mutating the Runtime at all. This suggests that I
should split the Runtime into two substructures: an immutable part
containing the local context, the builtins, the global context, etc, and
a mutable part which contains the readline Editor and the hole
information. The evaluation part can then be purely functional, while
hole-filling can be mutative.