Modify Parser to track the starting HoleId and the number of added holes.
I realized there was a bug when trying to fill a hole with an expression
containing another hole. The problem had to do with the fact that the
parser would always start indexing the HoleIds starting with zero, even
if the active Runtime had some already. This would end up with the new
holes (provided by the parser during hole-filling) to have HoleIds that
conflicted with the existing holes. And thus, filling a hole would
result in the Runtime confusing the two holes and clobbering values.

This commit fixes this issue by having the parser keep track of the
starting HoleId and the number of holes. The top-level parsing methods
now return pairs, indicating the parsed object and the number of holes
that parsed objet contains.

The new Runtime add_holes() method is called to register the given
number of new holes. (Holes are always assumed to be labeled with
HoleId 0 through N-1, where N is the number of holes.

Finally, the new hole_value() method queries the Runtime for the value
of the given hole.