Remove 'continue' command in hole-fill mode.
It was tedious to have to both 'fill' a value for a hole and then
'continue' to resume computation. I tried to think of a reason why this
should take two steps instead of one, but I couldn't. So I removed the
'continue' command and changed 'fill' to continue automatically.