Fix error in grammar for parsing types.
Previously, the parser was not properly parsing types which began with
a parenthesized type followed by an arrow. For example, the following
would fail to parse:

    (Nat -> Unit) -> Nat

The resolution was to create a new nonterminal in the grammar, TYPEPART,
(mirroring TERMPART). The production rules are now:

    TYPE := TYPEPART | TYPEPART -> TYPE
    TYPEPART := ATOM | ( TYPE )

Additionally, I fixed a bug where a type was being parsed incorrectly
due to not reversing the parsed parts of the type and not folding
properly.