Handle definitions with a Context, keeping in uniform with locals and builtins.
Previously, definitions were being treated very differently from
locals and builtins. They were being looked up by a separate
definition() method on Runtime. This was the case for a few reasons:

    * to facilitate recursive definitions
    * to relax the ordering of top-level bindings
    * to suppress evaluation (eg. println) that might happen during definition

However, upon thinking carefully about these situations, none of them
really mattered except the ordering. (Having order matter is
inconvenient, but not critical, and will have to be addressed separately
in the future anyway).

I renamed the definitions field on Runtime to definition_ctx and changed
it to a Context. The old definition() method has been removed. The
Runtime load() method was changed to simply extend the definition_ctx
with each new binding as it loads the file. Critically, the imports must
now be loaded before the current module's definitions can be evaluated.

Lastly, the Term::Var evaluation collapses significantly, amounting to
nothing more than an or_else() chain on the various Contexts in the
proper order.