Define InductiveTypeDefs.
In order to do typechecking for match statements, we need to keep a
record of the definitions for inductive types, like Nat and Bool. When
looking at the MatchArms of a match statement, we need the complete list
of constructor tags (CtorTags) that go together with the type of the
discriminee. And to apply the discriminee value to the matched arm, we
need to know the type of the matching arm's constructor.

This commit defines an InductiveTypeDef struct which captures this data.
The typedefs for simple builtin types (Nat, Bool, and Unit) are
constructed in the new builtin_inductive_typedefs() function. The list
of inductive typedefs loaded into the runtime are stored on a new field,
inductive_typedefs.