In a hole.
Create new enum hole::Command, which is the list of commands that the
user can perform while attempting to fill a hole.

I tried to think of as many spellings as I could for both "help" and
"quit", as it is needlessly frustrating when using a command-line tool
not to know how to exit or get help. I also added a cute bit of help
text in assets/help/hole.txt for when a user runs the help command.

To fill a hole, the user must explicitly write:

    > fill <expr>

This keeps the hole-filling process feeling more like an interactive
session (which it should be) rather than feeling like sudden user input.
To resume execution, the user may write:

    > continue

To abort the program, they may write:

    > abort

The value provided with fill is tracked, and calling fill again will
update the value. Calling continue without providing a value presents
an error message.