Prelude -> Builtins
When filling a hole, the local context displays all of the builtin
bindings (zero, succ, etc) and it's a bit overwheling to look at.

This commit creates a new context attached to the runtime called
builtin_ctx. All of the former "prelude" bindings have been moved to
this context. Evaluating a variable now consults this context. The order
goes:

    1) local context
    2) global bindings
    3) builtin context