Add type ascription and priliminary type checking.
This commit adds Token::As and Token::Colon and parses types and type
ascription. When loading a module, it typechecks all top-level bindings.
The Runtime now also manages TypeContexts for the builtins and for the
definitions.

The top-level types are given by the syntax:

    def add n m : Nat -> Nat -> Nat = ...

Some of the builtins have fudged types. I can't yet express universal
types, so Lists and Pairs aren't in yet.