Move readline editor into Runtime.
Save and load to a history file at:

    $HOME/.config/quail/history

This required changing eval and a few helper functions to take &mut
references to the Runtime rather than & references.

This also adds a new dependency dirs = "2.0.1" to get access to the
dirs::config_dir() function used to construct the config directory in
a platform-independent way.