Save values after holes have been filled.
Changes the evaluator so that when a hole is encountered, the runtime is
first consulted to see if the value has been filled in. If it has, then
that value is supplied. Otherwise, the user is prompted for a value to
fill the hole with, and that value is saved for later.

This prevents the erroneous situation where a single hole might
otherwise be filled in with multiple distinct values over the course of
the execution of a program. With these changes, once a hole is filled,
it will not need to be filled in again for the remainder of the program.