Add FilePathImportResolver and ChainedImportResolver.
These two ImportResolvers are used by main. The first,
FilePathImportResolver loads a module based on the exact filepath (as
opposed to appending ".ql" and searching in the examples/ directory. The
second, ChainedImportResolver, allows you to combine two ImportResolvers
which tries to resolve the import with the first, and then, if that
fails, it tries to resolve the import with the second.