Load builtin definitions from assets/types.txt.
I'm not ready to create syntax for these definitions, but at the same
time, building up the definitions in Rust is tedious and muddying to the
surrounding code. So I've created a small text file, assets/types.txt,
to contain the builtin inductive and coinductive type definitions using
a simple syntax.