Add force() and force_deep() to support thunk evaluation.
The force() function is used whenever a function is applied or a match
statement's discriminee is evaluated. This function will bring the value
to a kind of "head normal form", where the Value is anything other than
a Thunk, but the sub-values may still contain thunks (eg, under
constructors, under lambdas, in match arms).

The force_deep() function is needed to bring a value to normal form. In
particular, it's needed for inductive data types. It is called on a value
before sending it to a primitive.