Rename Unit/unit to Top/top and add Bot.
To support Bot, I needed to make a small change to the parser allowing a
match statement to have zero match arms. In that case, though, we can't
do a reverse lookup of the type. For now, we just assume that having no
match arms means you are trying to check against the Bot type.

I added a file example/trivial.ql demonstrating the Top and Bot types.