Add de Bruijn-style indexing notation to variables.
Normally, when creating a lambda abstraction or a let binding, the newly
introduced variable binding will shadow any previous variable in scope
by that same name. With this change, it will still be possible to refer
to shadowed variables, as illustrated in this example:

    fun x => (fun x => x$1)

Here, n$1 refers to the n variable n in the outer scope. Using this
notation, you can skip past `k` "copies" of the named variable by
writing the expression `x$k`.

The intent behind this change is that it opens up a new opportunity for
evaluation. Normally, it is problematic to perform Term-level
substitution using normal order because it can cause variable capture.
Here is an example:

    let m = 1 in
    (fun n m => n + m) m 2

This program reduces to the value 3 using eager normalization. However,
if we reduce this program by first applying `m` to the function
`fun n m => n + n`, we would get:

    let m = 1 in
    (fun m => m + m) 2

Which reduces to 4 instead. The reason is that the variable `m` has
been captured by the substitution into the function's body.

Normally, languages which reduce with Terms stipulate that you must
perform variable renaming first to avoid capture. This is fine when
presenting logics, but it becomes a burden to a computer to choose how
to do variable renaming in an automated fashion.

With this new syntax, however, we can canonically decide on a renaming
by appending `$1` to a freshly captured variable or, in the case a
variable `x$n` is captured, we can replace it with `x${n+1}`.